ml_project/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
click==7.1.2
colorama==0.4.4
dataclasses==0.8
importlib-metadata==4.0.1
iniconfig==1.1.1
joblib==1.0.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
typeguard==2.12.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
zipp==3.4.1